Refactor the installer if-else ladder
The module supplies two install methods that have been branches of an if-else
ladder in the serf::install class. Break each of these cases into its own
lower-level class that is then invoked according to the selected method. This
also allows for the module user to supply arbitrary install classes from
outside of the module, for example to build from source.